Refactor and reorganize project structure: - Moved `snake.html` to `Unused/` folder and updated references. - Extracted inline styles from `time_tracking.html` to `time_tracking_styles.css`. - Removed `about.html`. - Adjusted time tracking duration calculations. - Minor date formatting tweak in `script.js`.

+++ script.js
@@ -29,8 +29,8 @@ function showCurrentDate() {
     dateElement.id = 'current-date';
     
     const options = { 
-        weekday: 'long', 
-        year: 'numeric', 
+        weekday: 'short',
+        year: 'numeric',
         month: 'long', 
         day: 'numeric' 
     };
